feat(admin): new admin module with quarkus, micronaut and springboot integrations (#233)

+++ core/src/test/java/tech/illuin/pipeline/observer/PipelineTest.java
@@ -54,7 +54,7 @@ public class PipelineTest
             .findFirst().orElse(null));
         Assertions.assertEquals(
             counter.count(),
-            description.metrics().get(PIPELINE_RUN_SUCCESS_KEY.id()).values().get(counter.getId())
+            description.metrics().get(PIPELINE_RUN_SUCCESS_KEY.id()).values().get(counter.getId()).doubleValue()
         );
 
         Assertions.assertEquals("anonymous-init", description.init().id());
